[FEATURE-IMPROVEMENTS] LOG finished, check if the mobile phone has 9 digits, links to books and movies, delete JSON files and corrected user history

+++ LibraryApp/Business/LogService.cs
@@ -1,7 +1,7 @@
 using LibraryApp.Data;
 
 namespace LibraryApp.Business {
-    public class LogService : ILogService {
+    public class LogService {
         private readonly string _logFile;
         private int _errorCount;
 
@@ -13,20 +13,14 @@ namespace LibraryApp.Business {
         }
 
         public void LogError (string message) {
-            try {
-                _errorCount++;
-                string logMessage = $"ERROR {_errorCount}: {message}";
-                File.AppendAllText(_logFile, logMessage + Environment.NewLine);
-            }catch (Exception e) {
-                Console.WriteLine($"Error: {e.Message}");
-            }
+            _errorCount++;
+            string logMessage = $"ERROR {_errorCount}: {message}";
+            File.AppendAllText(_logFile, logMessage + Environment.NewLine);
         }
 
-        private void EndProgram(object sender, EventArgs e) {
-            try {
+        private void EndProgram(object? sender, EventArgs e) {
+            if (File.Exists(_logFile)) {
                 File.Delete(_logFile);
-            } catch (Exception ex) {
-                Console.WriteLine($"Error: {ex.Message}");
             }
         }
